+++ subnet_group.tf
@@ -1,6 +1,6 @@
 resource "aws_db_subnet_group" "database_subnet_group" {
   name        = "${var.component}-${var.deployment_identifier}"
-  description = "Subnet group for ${var.component} MySQL instance."
+  description = "Subnet group for ${var.component} MySQL database instance with deployment identifier ${var.deployment_identifier}."
   subnet_ids  = var.private_subnet_ids
 
   tags = {