+++ variables.tf
@@ -1,10 +1,13 @@
 variable "vpc_id" {
   description = "The ID of the VPC into which to deploy the database."
 }
+
+// TODO: rename this to subnet IDs since they could be public.
 variable "private_subnet_ids" {
   description = "The IDs of the private subnets to deploy the database into."
-  type = list(string)
+  type        = list(string)
 }
+// TODO: switch security group rules to be less specific about whether public or private.
 variable "private_network_cidr" {
   description = "The CIDR of the private network allowed access to the database."
 }
@@ -18,22 +21,37 @@ variable "deployment_identifier" {
 
 variable "database_instance_class" {
   description = "The instance type of the database instance."
-  default = "db.t2.micro"
+  default     = "db.t2.micro"
+}
+variable "storage_type" {
+  description = "The storage type to use for the database instance. One of \"standard\" (magnetic), \"gp2\" (general purpose SSD), \"gp3\" (general purpose SSD that needs iops independently) or \"io1\" (provisioned IOPS SSD). The default is \"standard\"."
+  default     = "standard"
+}
+variable "storage_iops" {
+  description = "The amount of provisioned IOPS. Can only be set when `storage_type` is \"io1\" or \"gp3\"."
+  type        = number
+  default     = null
 }
 variable "allocated_storage" {
   description = "The allocated storage in GBs."
-  default = 10
+  type        = number
+  default     = 10
+}
+variable "max_allocated_storage" {
+  description = "When configured, the upper limit to which Amazon RDS can automatically scale the storage of the DB instance. Configuring this will automatically ignore differences to `allocated_storage`. Must be greater than or equal to `allocated_storage` or 0 to disable Storage Autoscaling."
+  type        = number
+  default     = null
 }
 variable "database_version" {
   description = "The database version. If omitted, it lets Amazon decide."
-  default = ""
+  default     = ""
 }
 variable "database_family" {
   description = "The database family."
 }
 variable "database_port" {
   description = "The port the database listens on. 3306 by default."
-  default = "3306"
+  default     = "3306"
 }
 
 variable "database_name" {
@@ -48,9 +66,9 @@ variable "database_master_user_password" {
 
 variable "database_parameters" {
   description = "A list of parameters containing \"name\", \"value\" and \"apply_method\"."
-  type = list(object({
-    name = string
-    value = string
+  type        = list(object({
+    name         = string
+    value        = string
     apply_method = string
   }))
   default = []
@@ -58,10 +76,26 @@ variable "database_parameters" {
 
 variable "use_multiple_availability_zones" {
   description = "Whether or not to create a multi-availability zone database (\"yes\" or \"no\")."
-  default = "no"
+  default     = "no"
 }
 variable "use_encrypted_storage" {
   description = "Whether or not to use encrypted storage for the database (\"yes\" or \"no\")."
+  default     = "no"
+}
+variable "allow_public_access" {
+  description = "Whether or not to allow public access to the database (\"yes\" or \"no\")."
+  default     = "no"
+}
+variable "allow_major_version_upgrade" {
+  description = "Whether or not major version upgrades are allowed (\"yes\" or \"no\"). Changing this parameter does not result in an outage and the change is asynchronously applied as soon as possible. Defaults to \"no\"."
+  default     = "no"
+}
+variable "enable_automatic_minor_version_upgrade" {
+  description = "Whether or not minor engine upgrades will be applied automatically to the DB instance during the maintenance window (\"yes\" or \"no\"). Defaults to \"yes\"."
+  default     = "yes"
+}
+variable "enable_performance_insights" {
+  description = "Whether or not performance insights are enabled for the database (\"yes\" or \"no\"). Defaults to \"no\".."
   default = "no"
 }
 
@@ -72,13 +106,14 @@ variable "snapshot_identifier" {
 
 variable "backup_retention_period" {
   description = "The number of days to retain database backups."
-  default = 7
+  type        = number
+  default     = 7
 }
 variable "backup_window" {
   description = "The time window in which backups should take place."
-  default = "01:00-03:00"
+  default     = "01:00-03:00"
 }
 variable "maintenance_window" {
   description = "The time window in which maintenance should take place."
-  default = "mon:03:01-mon:05:00"
+  default     = "mon:03:01-mon:05:00"
 }